+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
-## 0.5.2.dev0 - next release
+## 0.5.2
 
-* TODO Fill this changes list while developing
+* Add trim option to marginal rate computation
+* Compute stop date of whole tax scale instead of a single bracket.
+* Use bracket stop date instead of legislation stop date inside a bracket.
+* Introduce reference periods
+* Fix roles_count bug : was set to 1 for scenarios vectorially defined
 
 ## 0.5.1